backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy>=2.0.36
pydantic>=2.10.0
httpx>=0.28.0
openai>=1.58.0
python-dotenv>=1.0.1
python-multipart>=0.0.12
slowapi>=0.1.9
aiosqlite>=0.20.0
typer>=0.15.1
rich>=13.9.4
anthropic>=0.42.0
google-generativeai>=0.8.3
